Moving utils to a separate file
